+++ internal/status/cpu.go
@@ -6,7 +6,7 @@ import (
 	"time"
 
 	"github.com/rechecked/rcagent/internal/config"
-	"github.com/shirou/gopsutil/v3/cpu"
+	"github.com/shirou/gopsutil/v4/cpu"
 )
 
 type CPUStatus struct {