+++ internal/status/docker.go
@@ -2,7 +2,7 @@ package status
 
 import (
 	"github.com/rechecked/rcagent/internal/config"
-	"github.com/shirou/gopsutil/v3/docker"
+	"github.com/shirou/gopsutil/v4/docker"
 )
 
 func HandleDocker(cv config.Values) interface{} {