+++ internal/status/services_windows.go
@@ -4,7 +4,7 @@
 package status
 
 import (
-	"github.com/shirou/gopsutil/v3/winservices"
+	"github.com/shirou/gopsutil/v4/winservices"
 	"golang.org/x/sys/windows"
 )